amazon-ai-chatbot/requirements.txt
# Core Dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0

# LLM Dependencies
llama-cpp-python>=0.2.0

# Optional: For better performance
# llama-cpp-python[server]>=0.2.0  # If you want to run as a server

# Data Processing
scikit-learn>=1.3.0

# Logging and Utilities
python-dateutil>=2.8.0

# Development Dependencies (optional)
# jupyter>=1.0.0
# matplotlib>=3.7.0
# seaborn>=0.12.0

# System Requirements Notes:
# - Python 3.8+ required
# - For Apple Silicon Macs: CMAKE_ARGS="-DLLAMA_METAL=on" pip install llama-cpp-python
# - For CUDA support: CMAKE_ARGS="-DLLAMA_CUBLAS=on" pip install llama-cpp-python
# - For CPU optimization: CMAKE_ARGS="-DLLAMA_BLAS=ON -DLLAMA_BLAS_VENDOR=OpenBLAS" pip install llama-cpp-python